requirements.txt
Flask
Flask-Cors
numpy
Pillow
tensorflow
gunicorn
# waitress
python-dotenv
